(patch-my-pc) Fixes Update Logic
The releases page we were scanning only deals with PMP Publisher at this
point, it seems, so the build was failing to find anything.

This Home Updater release notes page seems to be the right page for
the free updater, though it's worth noting that the download link has
changed since the 5.x release, and the 5.x release also only supports x64
Windows (see: https://patchmypc.com/faq/what-happens-now-that-patch-my-
pc-home-updater-version-4-is-deprecated/ for more details).